Fold SetContainerParent and SetContainerRelation hook (closes #3221815)
Imagine the following::

 class Babar(EntityType):
     country = Relation('Country', composite='object' ...)

 class Celeste(EntityType):
     country = Relation('Country', composite='object' ...)
     home = Relation('Home', composite='object' ....)


We have a situation where the `country` rtype is used from two entity
types, the first having country as sole upward link to the container
(hence Babar does not need a `container_parent` relation), the second
having two upward paths (country and home) to the container (hence
Celeste needs the `container_parent` relation.

To deal with this situation, we merge the `SetContainerParent` hook
into the `SetContainerRelation` hook and the later receives additonnal
rdef info to process from/to etypes in order not to try to stick a
`container_parent` where it does not belong.