[clone] provide a _fast_reserve_eid method to avoid one commit per eid
This is a major cost of the eid handling right now in cubicweb.

We monkeypatch the native source by anticipating the same future
cubicweb functionality that will allow to reserve an eid range in one
transaction.

Right now, we avoid the one-transaction-per-eid cost, but still pay a
call to the sequence generator for each eid, because postgres uses a
sequence that can't provide a contiguous integer range.

The monkeypatch should be dropped when the numrange sequence type is
made available in cubicweb through `.create_eid`.

See https://www.cubicweb.org/ticket/3526594

Related to  #3521652.