[clone] fix inlined relation handling
a) build set of rtypes that (strictly) crosses the border
b) build set of rtypes that have been already cloned

Both sets can be sent immediately to the attributes dict.

This is still insufficient because we cannot compute the opmtilal set
for b) nor do something interesting with rtypes who have some targets
inside and others outside the container, but it should at least be
correct.

We also add logs that will help decide that some inlined relations
are an anti-optimisation.

Related to #3521652.