[config] provide `.container_adapter` method to the configuration object
This will help automate the declaration and handling of the base
`ContainerProtocol` adapter (the one that provides
`.related_container` and `.parent`).

The method tries to provide *at most* one adapter per etype, globally.
We don't want silly SelectAmbiguity errors when having two containers
with some overlapping inner etypes (which is the case of `Diamond` and
`Mess` in test/).

It returns the adapter (for further use if wanted) and maintains a
cache for it.

The setup of entities/ is now quite simpler than it used to be.

The `test_adapter_without_container` test can now die.

Related to #3564655.