[config] config.container_adapter becomes a class method, handles ALL containers
The idea of one-adapter per container can't really fly because of the
overlap between containers.

This will be called in e.g. entities/... at a time when all containers have been
registered. Hence we can safely build one single adapter with the right selector
on it.

Also the registration is better made explicit.

Related to #3564655.