Rebuild the container relation for subgraphs being attached to the container
The new hook watches the `container_rtype` relation addition, so that when a
pre-existing subgraph of a container is attached to the container (i.e. the
`container_rtype` relation is added between the subgraph head and the
container) the AddContainerRelationOp is instanciated.

Turn the critical message into an informative warning message in
AddContainerRelationOp.

Closes to #3552734.

(Grafted from b74387f4abf2 and adapted to container configuration setup.)