Add ContainerConfig.structure and ContainerConfig.inner_rtypes methods
The first is a direct replacement of utils.container_static_structure which is
so deprecated. The second only yields non structural relations among two
entities which are both part of the container.

Use them together instead of now deprecated utils.container_rtypes_etypes (see
usage in entities.ContainerClone for instance).

Related to #3564655.