Make ContainerProtocol abstract
As a side effect (i.e. the adapter is not registered with the default yes()
selector), entities not in a container cannot be adapted to the
ContainerProtocol anymore (previously, for some mysterious reason, they were).
So adjust tests that relied on this "feature" and only ensure that entities
not in the container cannot be adapted.