Handle entities not related to their container in _set_container_parent
Fix a crash in `_set_container_parent` when `target` already has a
`container_parent` which is not related to a container (yet). So when adapting
`target` and `parent` to Container, both do not have any `related_container`,
hence the `AttributeError: 'NoneType' object has no attribute 'eid'`.

The new tests illustrates that by building the following entities:

        [ClownCabal]    [ClownCabal]
                   \    /
                  (members)
                    \  /
                   [Clown]

The bug is triggered when the second `members` relation is added, while the
`Clown` already has a `container_parent` (the first `ClownCabal`). By not
crashing, the `container_parent` relation is now updated to point to the last
added `parent` (i.e. the second `ClownCabal` in this test case).

This is more or less related to #3488993.

Closes #3702094.