[migration] Only sync container rtype in 2.5.0 migration
Previously almost all (or all?) relations were synced. Now only the container
rtype (the one for which default permissions changed) is synced.

The script loops on all entity types in the schema, detects if an entity is a
container thanks to the `container_config` class attribute of the entity
class, the container configuration holds the container rtype name, which is
synced.

Closes #3938384.