Config: provide the `.setup_rdefs_security` helper
The old `secutils.setup_container_rtypes_security` is left unchanged.

Since the container relation is now reflexive, we don't have to
provide near_container permission rules (which is a nice
simplification from the end-user side).

Having used rdefs from the start, the implementation is quite
simplified compared to the old version.

The config object also gains a public:

* `.border_rdefs` -> relations that cross the container border

Since defining a correct security policy for containers is an heavily
application-dependant issue, we make border_perms optional.

Related to #3564655.
Related to #3547361.