Config: tighten .border_rdefs
It did not properly account for subcontainers.

The proposed behaviour may still not be entirely satisfactory
since it allows for some rdefs to be shared by several containers.

However, this "border" case is easily handled with PERMS logic.

NOTE: setup_rdefs_permissions must *always* walk the rdefs, if only to
`resolve` the PERMS objects.

Related to #3589827.