Migration: distinguish between fs schema vs repo schema
In some odd applications, it may happen that:

* define_container is called on a repo schema (in which relations are
  already weaved)

* some container-controlled rdefs must be migrated

In this case, the add_relation_definition call will fail to perform
its job because both the fs and repo schema will already "know" the
rdefs (except that they lack an eid).

If we add one bit on the schema object to distinguish between fs and
repo schema, this situation can be avoided.