Views: automate uicfg hiding the container `crtype`
Either the primary view or the forms should not show this relation.

Provide a test to illustrate the purpose.

Note that the test does not test the "cubicweb-ctl start ..." case,
for which the code in views.py (uicfg post-reload setup_ui) is
necessary.

The test must be careful of the cubicweb version used. In 457efde98629
(cubicweb repository), the "add" action was extended to support the
'attributes' section.